# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpaceXStarlink API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Plan not found
  /checkout/session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionBody"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "400":
          description: Bad request
  /checkout/success:
    get:
      operationId: checkoutSuccess
      tags: [checkout]
      summary: Handle checkout success
      parameters:
        - name: session_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSuccessResponse"
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, cancelled, past_due, all]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionList"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get a subscription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Subscription not found
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update subscription status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionBody"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
components:
  schemas:
...
    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [residential, roam, business]
        speed:
          type: string
        priceMonthly:
          type: number
        features:
          type: array
          items:
            type: string
        stripePriceId:
          type: string
        active:
          type: boolean
        popular:
          type: boolean
        description:
          type: string
        hardwarePrice:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - speed
        - priceMonthly
        - features
        - active
        - popular
        - description
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        planId:
          type: integer
        planName:
          type: string
        planCategory:
          type: string
        priceMonthly:
          type: number
        stripeSubscriptionId:
          type: string
        stripeCustomerId:
          type: string
        status:
          type: string
          enum: [active, cancelled, past_due, trialing, incomplete]
        address:
          type: string
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - planId
        - planName
        - planCategory
        - priceMonthly
        - status
        - createdAt
    SubscriptionList:
      type: object
      properties:
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - subscriptions
        - total
        - page
        - limit
    UpdateSubscriptionBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, cancelled, past_due]
    CreateCheckoutSessionBody:
      type: object
      properties:
        planId:
          type: integer
        email:
          type: string
        name:
          type: string
        address:
          type: string
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - planId
        - email
        - name
    CheckoutSession:
      type: object
      properties:
        sessionId:
          type: string
        url:
          type: string
      required:
        - sessionId
        - url
    CheckoutSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        subscription:
          $ref: "#/components/schemas/Subscription"
      required:
        - success
...